fix: be jwt not my exception

See merge request s08-final/S08P31D202!66

+++ BE/skyeye/src/main/java/com/ssafy/skyeye/service/impl/BuildingServiceImpl.java
@@ -12,6 +12,7 @@ import com.ssafy.skyeye.repository.CrackRepository;
 import com.ssafy.skyeye.repository.UserRepository;
 import com.ssafy.skyeye.service.BuildingService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -41,6 +42,9 @@ public class BuildingServiceImpl implements BuildingService {
     public void registBuilding(BuildingRegistDto buildingRegistDto) {
         // TODO: JWT 본인인증 만약 아닐 시 405 Not Allowed
 
+        String jwtId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!jwtId.equals(buildingRegistDto.getUserId()) && !jwtId.equals("admin")) throw new ForbiddenException("본인 아이디가 아닙니다.");
+
         if(buildingRepository.existsByBuildingAddress(buildingRegistDto.getBuildingAddress())){
             throw new ForbiddenException("이미 있는 건물입니다.");
         }
@@ -59,6 +63,9 @@ public class BuildingServiceImpl implements BuildingService {
     @Override
     @Transactional
     public void updateBuilding(BuildingUpdateDto buildingUpdateDto) {
+        String jwtId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!jwtId.equals(buildingUpdateDto.getUserId()) && !jwtId.equals("admin")) throw new ForbiddenException("본인 아이디가 아닙니다.");
+
         Building building = getByBuildingId(buildingUpdateDto.getBuildingId());
         // TODO: JWT 본인인증 만약 아닐 시 405 Not Allowed
 
@@ -70,12 +77,19 @@ public class BuildingServiceImpl implements BuildingService {
     @Override
     public void deleteBuilding(Long buildingId) {
         // TODO: JWT 본인인증 만약 아닐 시 405 Not Allowed
+
+        String jwtId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!jwtId.equals(getByBuildingId(buildingId).getUserId().getUserId()) && !jwtId.equals("admin")) throw new ForbiddenException("본인 아이디가 아닙니다.");
+
         buildingRepository.delete(getByBuildingId(buildingId));
     }
 
     @Override
     public List<CrackDto> getBuildingByCrackId(Long buildingId) {
         // TODO: JWT 본인인증 만약 아닐 시 405 Not Allowed
+        String jwtId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!jwtId.equals(getByBuildingId(buildingId).getUserId().getUserId()) && !jwtId.equals("admin")) throw new ForbiddenException("본인 아이디가 아닙니다.");
+
         return crackRepository.findAll().stream()
                 .filter(crack -> crack.getBuildingId().getBuildingId().equals(buildingId))
                 .map(CrackDto :: entityToDto)
@@ -86,6 +100,9 @@ public class BuildingServiceImpl implements BuildingService {
     @Override
     public BuildingDto getBuilding(Long buildingId) {
         // TODO: JWT 본인인증 만약 아닐 시 405 Not Allowed
+        String jwtId = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!jwtId.equals(getByBuildingId(buildingId).getUserId().getUserId()) && !jwtId.equals("admin")) throw new ForbiddenException("본인 아이디가 아닙니다.");
+
         return BuildingDto.entityToDto(getByBuildingId(buildingId));
 
     }
